Dialog v2: move files, refaormat stories and add e2e tests
