+++ docs/content/drafts/Dialog.mdx
@@ -4,7 +4,7 @@ componentId: dialog
 status: Draft
 ---
 
-import data from '../../../src/Dialog/Dialog.docs.json'
+import data from '../../../src/Dialogv2/Dialog.docs.json'
 
 ```js
 import {Dialog} from '@primer/react/drafts'
